MultiROM: KillProcesses should not include recovery or lower pids
* Related to https://github.com/nkk71/android_bootable_recovery/commit/3cfa4d20fe06c0415665c343b36bcc1a8d4610d4

* If for some reason recovery (or lower pid, eg parent), has one of the fake
  mounts open/busy, we could accidentally cause a crash.

+++ multirom/Process.cpp
@@ -210,6 +210,10 @@ int Process::killProcessesWithOpenFiles(const char *path, int signal) {
             continue;
         }
 
+        // Let's not kill our own process
+        if (pid <= getpid())
+            continue;
+
         if (signal != 0) {
             LOGINFO("Sending %s to process %d", strsignal(signal), pid);
             kill(pid, signal);
